moved back and forward buttons to posts.utils.ts
